checkpoint: renames "core.ts" to "external.ts" in "/vue/Ref"
