+++ src/library/vue/Ref/exports.toolbox.ts
@@ -1,4 +1,4 @@
-export * from './core';
+export * from './external';
 
 export * from './get';
